requirements.txt
typer>=0.12.0
rich>=13.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
openai>=1.0.0
requests>=2.0.0
pytest>=8.0.0